small changes, a ton of error fixes, undo file flattening
