+++ Content/SkillPassives/FlameSage/Accelerant.cs
@@ -9,5 +9,5 @@ internal class Accelerant(SkillTree tree) : SkillPassive(tree)
 	public const float DamageIncrease = 0.15f;
 	public const float LifeLoss = 0.05f;
 
-	public override string DisplayTooltip => string.Format(base.DisplayTooltip, MathUtils.Percent(LifeLoss), MathUtils.Percent(DamageIncrease));
+	public override object[] TooltipArguments => [MathUtils.Percent(LifeLoss), MathUtils.Percent(DamageIncrease)];
 }
