+++ Content/SkillPassives/FlameSage/EnduringFlame.cs
@@ -7,5 +7,5 @@ namespace PathOfTerraria.Content.SkillPassives.FlameSage;
 internal class EnduringFlame(SkillTree tree) : SkillPassive(tree)
 {
 	public const float DamageBonus = 0.02f;
-	public override string DisplayTooltip => string.Format(base.DisplayTooltip, MathUtils.Percent(DamageBonus));
+	public override object[] TooltipArguments => [MathUtils.Percent(DamageBonus)];
 }
