+++ Content/SkillPassives/FlameSage/FlameWard.cs
@@ -7,5 +7,5 @@ namespace PathOfTerraria.Content.SkillPassives.FlameSage;
 internal class FlameWard(SkillTree tree) : SkillPassive(tree)
 {
 	public const float ResistanceIncrease = 0.2f;
-	public override string DisplayTooltip => string.Format(base.DisplayTooltip, MathUtils.Percent(ResistanceIncrease));
+	public override object[] TooltipArguments => [MathUtils.Percent(ResistanceIncrease)];
 }
