+++ Content/SkillPassives/FlameSage/ImprovedWarding.cs
@@ -7,7 +7,7 @@ namespace PathOfTerraria.Content.SkillPassives.FlameSage;
 internal class ImprovedWarding : SkillPassive
 {
 	public const float ResistanceIncrease = 0.05f;
-	public override string DisplayTooltip => string.Format(base.DisplayTooltip, MathUtils.Percent(ResistanceIncrease));
+	public override object[] TooltipArguments => [MathUtils.Percent(ResistanceIncrease)];
 
 	public ImprovedWarding(SkillTree tree) : base(tree)
 	{
