+++ Content/SkillPassives/FlameSage/IntenseHeat.cs
@@ -7,5 +7,5 @@ namespace PathOfTerraria.Content.SkillPassives.FlameSage;
 internal class IntenseHeat(SkillTree tree) : SkillPassive(tree)
 {
 	public const float DamageIncrease = 0.2f;
-	public override string DisplayTooltip => string.Format(base.DisplayTooltip, MathUtils.Percent(DamageIncrease));
+	public override object[] TooltipArguments => [MathUtils.Percent(DamageIncrease)];
 }
