+++ Content/SkillPassives/FlameSage/MeltingPoint.cs
@@ -9,5 +9,5 @@ internal class MeltingPoint(SkillTree tree) : SkillPassive(tree)
 	public const float DamageBonus = 0.2f;
 	public const float Seconds = 3f;
 
-	public override string DisplayTooltip => string.Format(base.DisplayTooltip, MathUtils.Percent(DamageBonus), Seconds);
+	public override object[] TooltipArguments => [MathUtils.Percent(DamageBonus), Seconds];
 }
