+++ Content/SkillPassives/FlameSage/SlowBurn.cs
@@ -6,5 +6,5 @@ namespace PathOfTerraria.Content.SkillPassives.FlameSage;
 internal class SlowBurn(SkillTree tree) : SkillPassive(tree)
 {
 	public const float Seconds = 4;
-	public override string DisplayTooltip => string.Format(base.DisplayTooltip, Seconds);
+	public override object[] TooltipArguments => [Seconds];
 }
