+++ Content/SkillPassives/NovaTree/Combustive.cs
@@ -1,7 +1,6 @@
 ﻿using PathOfTerraria.Common.Mechanics;
 using PathOfTerraria.Common.Projectiles;
 using PathOfTerraria.Common.Systems.Skills;
-using PathOfTerraria.Common.Utilities;
 using PathOfTerraria.Content.Skills.Magic;
 using Terraria.GameContent.Drawing;
 using Terraria.ID;
@@ -49,45 +48,3 @@ internal class Combustive(SkillTree tree) : SkillPassive(tree)
 		}
 	}
 }
-
-internal class ConcurrentBlasts(SkillTree tree) : SkillPassive(tree)
-{
-	/// <summary> A damage multiplier applied to targets based on <see cref="ConcurrentNPC.Vulnerable"/>. </summary>
-	public const float BonusDamage = 1.15f;
-	public override object[] TooltipArguments => [MathUtils.Percent(BonusDamage - 1)];
-}
-
-internal class ScorchingTouch(SkillTree tree) : SkillPassive(tree)
-{
-	internal class FieryPatch : SkillProjectile<Nova>
-	{
-		public override string Texture => "Terraria/Images/NPC_0";
-
-		public override void SetDefaults()
-		{
-			Projectile.friendly = true;
-			Projectile.Size = new Vector2(20);
-			Projectile.timeLeft = 180;
-			Projectile.penetrate = -1;
-			Projectile.aiStyle = -1;
-			Projectile.tileCollide = false;
-			Projectile.ignoreWater = true;
-		}
-
-		public override void AI()
-		{
-			if (Main.rand.NextBool(10))
-			{
-				Vector2 position = Projectile.position + new Vector2(Projectile.width * Main.rand.NextFloat(), 4);
-				ParticleOrchestrator.SpawnParticlesDirect(ParticleOrchestraType.WallOfFleshGoatMountFlames, new() { PositionInWorld = position });
-			}
-		}
-	}
-}
-
-internal class ThunderClaps(SkillTree tree) : SkillPassive(tree)
-{
-	public override object[] TooltipArguments => ["5"];
-}
-
-internal class VolatileNova(SkillTree tree) : SkillPassive(tree);
