+++ Content/SkillPassives/NovaTree/ThunderClaps.cs
@@ -1,0 +1,9 @@
+﻿using PathOfTerraria.Common.Mechanics;
+using PathOfTerraria.Common.Systems.Skills;
+
+namespace PathOfTerraria.Content.SkillPassives.NovaTree;
+
+internal class ThunderClaps(SkillTree tree) : SkillPassive(tree)
+{
+	public override object[] TooltipArguments => ["5"];
+}
