+++ Content/SkillPassives/NovaTree/VolatileNova.cs
@@ -1,0 +1,6 @@
+﻿using PathOfTerraria.Common.Mechanics;
+using PathOfTerraria.Common.Systems.Skills;
+
+namespace PathOfTerraria.Content.SkillPassives.NovaTree;
+
+internal class VolatileNova(SkillTree tree) : SkillPassive(tree);
